requirements.txt
numpy
Flask
gunicorn
sklearn
pandas
Pyyaml
-e .